Refactor & update `stm.cabal` to `cabal-version>=1.10`
As the current code fails to build with GHC 6.10, I've updated the Cabal
meta-data to declare support for GHC>=6.12 and dropped support for
base3. This also moves the changelog from the Cabal description field
into a separate `changelog` file and adds a few entries.

The code compiles warning-free for all tested GHC versions.

Signed-off-by: Herbert Valerio Riedel <hvr@gnu.org>