Make `signalTSem` resilient against `Int` overflows
This is accomplished by switching the internal implementation to using
an 'Integer' value for the internal counter value which can never
overflow; the small computational overhead is assumed be neglectable
in the context of STM transaction overheads.

`waitTSem` was never affected as it already blocks when the counter
becomes non-positive.